+++ numerapi/cli.py
@@ -112,6 +112,16 @@ def submission_filenames(round_num=None, tournament=None):
 
 
 @cli.command()
+@click.option('--tournament', type=int, default=1,
+              help='ID of the tournament, defaults to 1')
+@click.option('--round_num', type=int, default=0,
+              help='round number, defaults to the current round')
+def staking_cutoff(round_num, tournament):
+    """Compute staking cutoff for the given round and tournament."""
+    click.echo(prettify(napi.get_staking_cutoff(round_num, tournament)))
+
+
+@cli.command()
 @click.option('--tournament', default=1,
               help='The ID of the tournament, defaults to 1')
 @click.option('--hours', default=24,